requirements.txt
homeassistant~=2023.12.1
async-timeout~=4.0.3
# autopep8~=2.0.4
ruff~=0.1.7
google-generativeai~=0.3.2
google-cloud-aiplatform~=1.38.1